+++ CHANGELOG.rst
@@ -3,7 +3,14 @@
 Changelog
 =========
 
-Version 0.1.0 - Unreleased
+Version 0.2.0 - Unreleased
 **************************
 
-* First commit.
+Added working version of requirement analyzer, parser and package model with
+commandline and tests.
+
+
+Version 0.1.0 - Not released
+****************************
+
+First commit with proof of concept with Libraries.io, requests and packaging.